+++ packages/react-core/src/components/atoms/Badge/index.ts
@@ -1,0 +1,1 @@
+export { default as Badge, BadgeProps } from './Badge';
