+++ packages/react-core/src/components/atoms/Text/Text.stories.tsx
