+++ packages/react-core/src/index.ts
@@ -1,16 +1,16 @@
-export { default as Button } from './components/atoms/button/Button';
-export * from './components/atoms/button/Button';
-export { default as Text } from './components/atoms/text/Text';
-export * from './components/atoms/text/Text';
+export { default as Button } from './components/atoms/Button/Button';
+export * from './components/atoms/Button/Button';
+export { default as Text } from './components/atoms/Text/Text';
+export * from './components/atoms/Text/Text';
 export {
   default as ThemeProvider,
   ThemeProp,
 } from './components/styles/ThemeProvider';
-export { default as Input, InputProps } from './components/atoms/input/Input';
+export { default as Input, InputProps } from './components/atoms/Input/Input';
 export * from './components/styles/definitions';
-export { Paper } from './components/atoms/paper/Paper';
-export { Card } from './components/atoms/card/Card';
-export { Badge } from './components/atoms/badge/Badge';
-export { Switch } from './components/atoms/switch/Switch';
-export { RadioButton } from './components/atoms/radioButton/RadioButton';
-export { CheckBox } from './components/atoms/checkbox/CheckBox';
+export { Paper } from './components/atoms/Paper/Paper';
+export { Card } from './components/atoms/Card/Card';
+export { default as Badge, BadgeProps } from './components/atoms/Badge/Badge';
+export { Switch } from './components/atoms/Switch/Switch';
+export { RadioButton } from './components/atoms/RadioButton/RadioButton';
+export { CheckBox } from './components/atoms/CheckBox/CheckBox';
